feat: retro team share ui (#22)
* feat: retro team share ui

* feat: retro invitation link dialog

* feat: invitation accept flow

* feat: team user join/leave handling

* chore: lint bupmed

* chore: lint bump

+++ apps/web/src/views/team_edit/TeamEditView.tsx
@@ -1,64 +1,36 @@
-import { AnimatePresence } from "framer-motion";
 import type React from "react";
 import { useEffect, useState } from "react";
 import { Navigate, useNavigate, useParams } from "react-router";
 import { toast } from "react-toastify";
-import type { InviteResponse } from "shared/model/invite/Invite.response";
 import type { TeamRequest } from "shared/model/team/team.request";
-import type { UserInTeamResponse } from "shared/model/user/user.response";
-import {
-  deleteTeam,
-  editTeam,
-  getInvitesByTeamId,
-  getTeamById,
-} from "../../api/Team.service";
-import { getUsersByTeamId } from "../../api/User.service";
+import { deleteTeam, editTeam } from "../../api/Team.service";
 import { ProgressBar } from "../../component/atoms/progress_bar/ProgressBar";
-import { ConfirmDialog } from "../../component/molecules/confirm_dialog/ConfirmDialog";
 import { AnimatedBackground } from "../../component/organisms/animated_background/AnimatedBackground";
 import { TeamForm } from "../../component/organisms/forms/TeamForm";
 import Navbar from "../../component/organisms/navbar/Navbar";
+import { useConfirm } from "../../context/confirm/ConfirmContext.hook";
+import { useTeamData } from "../../context/useTeamData";
 import { useUser } from "../../context/user/UserContext.hook";
 
-const TeamEditView: React.FC = () => {
+export const TeamEditView: React.FC = () => {
   const { teamId } = useParams<{ teamId: string }>();
   const { user, refreshUser } = useUser();
   if (!teamId || !user) {
     return <Navigate to={"/"} />;
   }
   const navigate = useNavigate();
+  const { showConfirm } = useConfirm();
+  const { team: initialTeam } = useTeamData(teamId);
+
   const [team, setTeam] = useState<TeamRequest | null>(null);
-  const [confirmOpen, setConfirmOpen] = useState(false);
 
   useEffect(() => {
-    const waitForResult = async () => {
-      const teamResponse = await getTeamById(teamId);
-
-      const users = await getUsersByTeamId(teamId).then((users) =>
-        users.filter((elem: UserInTeamResponse) => {
-          return elem.email !== user?.email;
-        }),
-      );
-
-      const invites = await getInvitesByTeamId(teamId).then((data) =>
-        data.map((invite: InviteResponse) => ({
-          email: invite.email,
-          role: invite.role,
-        })),
-      );
-
-      setTeam({
-        name: teamResponse.name,
-        users: [...users, ...invites],
-      });
-    };
-
-    waitForResult();
-  }, []);
+    setTeam(initialTeam);
+  }, [initialTeam]);
 
   const onSubmit = (team: TeamRequest) => {
     editTeam(teamId, team)
-      .then((response) => {
+      .then(() => {
         refreshUser().then(() => {
           navigate("/");
         });
@@ -72,20 +44,26 @@ const TeamEditView: React.FC = () => {
       });
   };
 
-  const onDelete = () => {
-    deleteTeam(teamId)
-      .then(() => {
-        refreshUser().then(() => {
-          navigate("/");
-        });
-      })
-      .then(() => {
-        toast.success("Zespół usunięto");
-      })
-      .catch((e) => {
-        console.log(e);
-        toast.error("Wystąpił błąd");
-      });
+  const onDeleteTeamPress = () => {
+    showConfirm({
+      title: "Usunięcie zespołu",
+      content: `Czy na pewno chcesz usunąć zespół ${team?.name ?? ""}?`,
+      onConfirmed: () => {
+        deleteTeam(teamId)
+          .then(() => {
+            refreshUser().then(() => {
+              navigate("/");
+            });
+          })
+          .then(() => {
+            toast.success("Zespół usunięto");
+          })
+          .catch((e) => {
+            console.log(e);
+            toast.error("Wystąpił błąd");
+          });
+      },
+    });
   };
 
   return (
@@ -99,21 +77,10 @@ const TeamEditView: React.FC = () => {
           </div>
         )}
 
-        <AnimatePresence>
-          {confirmOpen && (
-            <ConfirmDialog
-              title={"Usunięcie zespołu"}
-              content={`Czy na pewno chcesz usunąć zespół ${team?.name ?? ""}?`}
-              onConfirmed={onDelete}
-              onDismiss={() => setConfirmOpen(false)}
-            />
-          )}
-        </AnimatePresence>
-
         {team && (
           <TeamForm
             onSubmit={onSubmit}
-            onDelete={() => setConfirmOpen(true)}
+            onDelete={onDeleteTeamPress}
             team={team}
             deletable
           />
@@ -122,5 +89,3 @@ const TeamEditView: React.FC = () => {
     </>
   );
 };
-
-export default TeamEditView;
